Override the linkCalls() method for python
